// src/types.ts
import { AdTypes, PropertyTypes, CarBrands, ServiceTypes } from './constants';

export type FormValues = {
  // Общие поля
  name: string; // Название объявления
  description: string; // Описание объявления
  location: string; // Локация объявления
  type: (typeof AdTypes)[number]; // Тип объявления (обязательно)
  image?: string; // Ссылка на картинку (необязательное)

  // Дополнительные поля в зависимости от типа объявления

  // Недвижимость
  propertyType?: (typeof PropertyTypes)[number]; // Тип недвижимости (обязательно для "Недвижимость")
  area?: number; // Площадь в кв. метрах (обязательно для "Недвижимость")
  rooms?: number; // Количество комнат (обязательно для "Недвижимость")
  price?: number; // Цена в рублях (обязательно для "Недвижимость")

  // Авто
  brand?: (typeof CarBrands)[number]; // Марка автомобиля (обязательно для "Авто")
  model?: string; // Модель автомобиля (обязательно для "Авто")
  year?: number; // Год выпуска автомобиля (обязательно для "Авто")
  mileage?: number; // Пробег автомобиля в км (необязательно для "Авто")

  // Услуги
  serviceType?: (typeof ServiceTypes)[number]; // Тип услуги (обязательно для "Услуги")
  experience?: number; // Опыт работы в годах (обязательно для "Услуги")
  cost?: number; // Стоимость услуги в рублях (обязательно для "Услуги")
  workSchedule?: string; // График работы (необязательно для "Услуги")
};

export type Ad = {
  id: number; // Уникальный идентификатор объявления
  name: string; // Название объявления
  description: string; // Описание объявления
  location: string; // Локация
  type: 'Недвижимость' | 'Авто' | 'Услуги'; // Тип объявления (Категория)
  image?: string; // Ссылка на изображение (опционально)

  // В зависимости от категории
  propertyType?: string; // Тип недвижимости (если "Недвижимость")
  area?: number; // Площадь в квадратных метрах (если "Недвижимость")
  rooms?: number; // Количество комнат (если "Недвижимость")
  price?: number; // Цена (если "Недвижимость")

  brand?: string; // Марка автомобиля (если "Авто")
  model?: string; // Модель автомобиля (если "Авто")
  year?: number; // Год выпуска (если "Авто")
  mileage?: number; // Пробег (если "Авто")

  serviceType?: string; // Тип услуги (если "Услуги")
  experience?: number; // Опыт работы в годах (если "Услуги")
  cost?: number; // Стоимость услуги в рублях (если "Услуги")
  workSchedule?: string; // График работы (если "Услуги")
};